+++ CHANGELOG.md
@@ -3,4 +3,8 @@
 
 <small>[Compare with latest](https://github.com/cxpsemea/ts-inventory/compare/7a963650840f52d669c612f991f8bcef7d9aec8a...HEAD)</small>
 
+### Chore
+
+- sast inventory ([b0b9f49](https://github.com/cxpsemea/ts-inventory/commit/b0b9f4915dff8cddcfd82df3987599dc7c2ead4e) by jcostacx).
+
 <!-- insertion marker -->